# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AndMI API/Open API System",
        "description": "Andijon Mashinasozlik Instituti Ochiq API tizimi",
        "contact": {
            "email": "andmiatmdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/times": {
            "get": {
                "tags": [
                    "Временные интервалы"
                ],
                "summary": "Получение списка всех временных интервалов",
                "operationId": "4fa9086cab17eae107a3a8b18764650c",
                "responses": {
                    "200": {
                        "description": "Список временных интервалов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Рабочее время"
                                            },
                                            "start": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "08:00:00"
                                            },
                                            "end": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "17:00:00"
                                            },
                                            "monday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "tuesday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "wednesday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "thursday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "friday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "saturday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "sunday": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Временные интервалы"
                ],
                "summary": "Создание временного интервала",
                "operationId": "d181bc607c0e5e26ffd2940a9b0beda6",
                "requestBody": {
                    "description": "Данные для создания временного интервала",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start",
                                    "end",
                                    "monday",
                                    "tuesday",
                                    "wednesday",
                                    "thursday",
                                    "friday",
                                    "saturday",
                                    "sunday"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Рабочее время"
                                    },
                                    "start": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00:00"
                                    },
                                    "end": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00:00"
                                    },
                                    "monday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "tuesday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "wednesday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "thursday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "friday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "saturday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "sunday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Временный интервал успешно создан"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/times/{id}": {
            "get": {
                "tags": [
                    "Временные интервалы"
                ],
                "summary": "Получение временного интервала",
                "operationId": "05b4be7790d60c187be7fc0f93e7423e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID временного интервала",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Временный интервал найден"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Временный интервал не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Временные интервалы"
                ],
                "summary": "Обновление временного интервала",
                "operationId": "e0afaf3c23953b62885c435ee5dc3281",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID временного интервала",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления временного интервала",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "start",
                                    "end",
                                    "monday",
                                    "tuesday",
                                    "wednesday",
                                    "thursday",
                                    "friday",
                                    "saturday",
                                    "sunday"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Рабочее время"
                                    },
                                    "start": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00:00"
                                    },
                                    "end": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00:00"
                                    },
                                    "monday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "tuesday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "wednesday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "thursday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "friday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "saturday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "sunday": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Временный интервал успешно обновлен"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Временный интервал не найден"
                    },
                    "422": {
                        "description": "Некорректные данные"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Временные интервалы"
                ],
                "summary": "Удаление временного интервала",
                "operationId": "bf2856573ef28f2865c4b8320485b0ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID временного интервала",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Временный интервал успешно удален"
                    },
                    "401": {
                        "description": "Неавторизованный запрос"
                    },
                    "404": {
                        "description": "Временный интервал не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token orqali autentifikatsiyadan o'tish",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}